# mvcc: use buffer bytes to encode consistent index