```
go test -v -run=xxx -benchmem -bench BenchmarkStoreTxnPut

Before
BenchmarkStoreTxnPut-8	  200000	     11221 ns/op	    1476 B/op	      13 allocs/op

After 
BenchmarkStoreTxnPut-8	  200000	     10619 ns/op	    1463 B/op	      12 allocs/op
```

If it's not worth this, just remove the TODO?