@heyitsanthony 2016-05-18
I think it would be faster to have this temp buffer as a field in the store struct instead of going through a sync.Pool since saveIndex is already serialized on s.mu.Lock