@gyuho 2016-05-18
Just changed it to use bytes buffers inside store, as @heyitsanthony suggested, and it's simpler and still saving us a few bytes:

```
BEFORE:  200000	     11221 ns/op	    1476 B/op	      13 allocs/op
AFTER :  200000	     10442 ns/op	    1464 B/op	      12 allocs/op
```

PTAL.

Thanks.