@gyuho 2016-05-18
go test -v -run=xxx -benchmem -bench=BenchmarkStore

```
# BEFORE
BenchmarkStorePut-8   	  200000	     10749 ns/op	    1472 B/op	      13 allocs/op
BenchmarkStoreTxnPut-8	  200000	     10795 ns/op	    1471 B/op	      13 allocs/op

# AFTER
BenchmarkStorePut-8   	  200000	     11036 ns/op	    1467 B/op	      12 allocs/op
BenchmarkStoreTxnPut-8	  200000	     10326 ns/op	    1462 B/op	      12 allocs/op
```

Merging in greens.

Thanks!